+++ src/kanban/README.md
@@ -1,0 +1,38 @@
+# Практическая работа - Kanban.
+
+## Функционал
+
+1. Выпадающий список, который появляется при клике на блок в правом верхнем углу - контейнер с аватаром пользователя и стрелочкой.
+2. Канбан доска содержит 4 блока задач:
+    - Backlog(задачи, которые требуют уточнения, перед тем, как их брать в работу).
+    - Ready(задачи, которые могут быть взяты в работу).
+    - In progress (задачи, которые уже в работе).
+    - Finished (законченные задачи).
+3. Добавление новых/существующих задач в Backlog/Ready/In progress/Finished блоки задач.
+
+## Файловая структура
+```
+.
+├── src
+|   ├── css
+|   |   └── index.css
+|   ├── img
+|   |   ├── avatar.svg
+|   |   ├── dots.svg
+|   |   ├── menuGamb.png
+|   |   └── plus.svg
+|   └── js
+|       ├── board.js
+|       ├── data.js
+|       ├── header.js
+|       └── index.js
+├── index.html
+└── README.md
+```
+
+## Запуск проекта
+<span style="color:red">«!» ВАЖНО:</span> Используйте локальный сервер, чтобы поддерживать загрузку JS модулей!
+
+Для Visual Code можно использовать [Live Server](https://marketplace.visualstudio.com/items?itemName=ritwickdey.LiveServer):
+1. Установить плагин.
+2. Нажать Go Live (правый нижний угол VSCode), чтобы запустить сервер.
