+++ src/kanban/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="utf-8">
     <title>Kanban</title>
-    <link rel="stylesheet" href="task.css">
+    <link rel="stylesheet" href="./src/css/index.css">
 </head>
 
 <body>
@@ -12,7 +12,7 @@
         <div class="leftBlock">
             <div class="menu">
                 <button class="buttonMenu">
-                    <img class="imgMenu hoverAction" src="./imgKanban/menuGamb.png" alt="menu">
+                    <img class="imgMenu hoverAction" src="./src/img/menuGamb.png" alt="menu">
                 </button>
             </div>
             <div class="heading">
@@ -21,13 +21,13 @@
         </div>
         <div class="rightBlock">
             <button class="buttonCreateNewList hoverAction">
-                <img class="imgCreateNewList" src="./imgKanban/plus.svg" alt="Create list">
+                <img class="imgCreateNewList" src="./src/img/plus.svg" alt="Create list">
                 <span class="spanCreateNewList">Create new list</span>
             </button>
             <div class="separator"></div>
             <div class="avatar">
                 <button id="buttonDropDown" class="buttonUserMenu hoverAction arrowDown">
-                    <img class="imgAvatar borderAvatar" src="./imgKanban/avatar.svg" alt="avatar">
+                    <img class="imgAvatar borderAvatar" src="./src/img/avatar.svg" alt="avatar">
                 </button>
                 <nav id="dropDownUserMenu" class="hidden dropDownContent">
                     <ul>
@@ -57,7 +57,7 @@
             <span class="spanFooter">&ensp; &lt; NAME &gt;</span>
         </div>
     </footer>
-    <script src="task.js"></script>
+    <script src="./src/js/index.js" type="module"></script>
 </body>
 
 </html>
