+++ src/kanban/src/js/data.js
@@ -1,0 +1,68 @@
+export const defaultData = [
+    {
+        id: 5000,
+        title: 'Backlog',
+        issues: [
+            {
+                id: 1588614113700,
+                name: 'Sprint bugfix'
+            },
+            {
+                id: 1588614113701,
+                name: 'Login page – performance issues'
+            }
+        ]
+    }, {
+        id: 5001,
+        title: 'Ready',
+        issues: [
+            {
+                id: 1588614113702,
+                name: 'Shop bug1'
+            },
+            {
+                id: 1588614113703,
+                name: 'Shop bug2'
+            },
+            {
+                id: 1588614113704,
+                name: 'Shop bug3'
+            },
+            {
+                id: 1588614113706,
+                name: 'Shop bug4'
+            },
+            {
+                id: 1588614113707,
+                name: 'Shop bug5'
+            },
+            {
+                id: 1588614113708,
+                name: 'Shop bug6'
+            },
+            {
+                id: 1588614113709,
+                name: 'Shop bug7'
+            }
+
+        ]
+    }, {
+        id: 5500,
+        title: 'In Progress',
+        issues: [
+            {
+                id: 1588614113710,
+                name: 'Auth bugfix'
+            }
+        ]
+    }, {
+        id: 5502,
+        title: 'Finished',
+        issues: [
+            {
+                id: 1588614113711,
+                name: 'Main page bugfix'
+            }
+        ]
+    }
+];
